[FIX] mrp_account_enterprise: add fallback on reporting
This commit fixes a small bug on sql. Removing unused
join with company_id, avoid to add unnused lines. Use
the fields_id instead of ir_prop.name and add fallback
options for ir_prop. If the res_id is not found on the
database, then we should user the ir_prop without a res_id
or without company_id.

To test it, just open the production analysis, create
a BOM for it and check if the expected fields are satisfied.

closes odoo/enterprise#68327

X-original-commit: e8130529f8dcd7dfd613834f0cc482d0976c3c77
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>