[FIX] planning: avoid writing on field when cursor is readonly
Before this commit, `_compute_past_shift` method will compute the `is_past`
but also set `request_to_switch` field to False when the shift is in the
past and `request_to_switch` was truely. The problem is the `_compute_past_shift`
could be only called to read `is_past` field and so the cursor is
sometimes in readonly to only allow SQL queries to fetch data.

This commit makes sure the `_compute_past_shift` method will only alter
`is_past` field and alter `request_to_switch` inside the method that
could be called when `request_to_switch` is truly for the shift contains
in self. A cron is not added to make sure the request_to_switch for the
shifts in the past will be set to False because the actual code is
sufficient since we usually check if the shift is in the past before
allowing the action or displaying the button.

runbot-66162
task-4276516

closes odoo/enterprise#77701

X-original-commit: 04e44ee7a61dc31aa5355209793a9a4d474a1f70
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>