[FIX] account_reports: fix getting account report data without payment account id
Currently, an exception is generated when the user tries to open
the "Cash Flow Statement" report.

Steps to produce an error:
- install 'Accounting Reports'
- Open Accounting > Configuration > Accounting > Journals
- Now delete all journals except "Miscellaneous Operations"
- change type of "Miscellaneous Operations" to sales > sale
- Click Reporting > click "Cash Flow Statement"  >>> exception occurs

error:
```
SyntaxError
syntax error at or near ")"
LINE 28: ...            AND account_move_line.account_id NOT IN ((), ())
```

This is because at the time of query building for getting liquidity moves,
it gets an empty tuple from line [1] as we do not have any journals with
the type 'bank', 'cash', or 'general'.

This commit will fix the above issue by returning empty report_data
when payment_account_ids is empty.

[1] - https://github.com/odoo/enterprise/blob/8fa01948610809d191b75260772782e25ea3a737/account_reports/models/account_cash_flow_report.py#L195

sentry-5555757610

closes odoo/enterprise#69798

X-original-commit: f29a8f1e76b9d8bab46fe195f604b7950752765d
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: ANSARI MAHAMADASIF (maan) <maan@odoo.com>