[FIX] l10n_nl_reports_sbr_status_info: remove use of constraint
The constraint was failing because when uploading the certificate,
password and key through the settings, the values are written one by one
on the company, which made it fail in case the password is written first
(the constraint would test the password on empty values and throw an
error). As a certificate or private key can be un-crypted at the upload,
we can't be sure a password will be set up too, so, when checking the
certifcate and key, the password might be absent either because there is
none or because it wasn't written on the company yet (but we can't know).

We thus remove the constraint and add the check directly at the line for
which it was created.

The groups right is not useful anymore as the flow uses sudo to modify
it outside of the settings.

closes odoo/enterprise#68577

X-original-commit: 8ee49599a8e937e91c148163ebdf2b2d21503503
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>