[REF] html_editor: use the same api as the powerbox for the toolbar
The purpose of this commit is to modify the api used to define toolbar
elements so that it is like the powerbox api.

We're therefore going to replace toolbarGroup by toolbarCategory/toolbarCommands.
These are the equivalents of powerboxCategory and powerboxCommands.

Thanks to this new API, a plugin can add a command to an existing category.

closes odoo/enterprise#67932

Related: odoo/odoo#175507
Signed-off-by: Nicolas Bayet (nby) <nby@odoo.com>