[IMP] stock_barcode: split RFID
When using RFID scanner, it can happen multiple RFID to be send at once
as a single string. When that happens, the barcode parser cannot handle
it, giving an error.
To avoid that, the `BarcodeModel` will split the barcode if it's
actually multiple RFID.

task-4089638

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>