[IMP] stock_barcode: multi barcode better handling
This commit improves the Barcode app in multiple way to handle the scan
of multiple barcodes at once (usefull when scanning a package content or
a lot of RFID in a little time range.)

Single RPC for all barcodes
===========================

Before this commit, when multiple barcodes are scanned at once (barcodes
separated by the separator char/regex, or multiple RFID send as one
string for example), each barcode is processed individually, which means we will do
the follow steps for each barcode:

1. The nomenclature parses the barcode;
2. We use the barcode parsed data to get the record's data (which causes a RPC if
   the record data are not in the JS cache);
3. Then, we use those data to update the application itself (e.g.: add a new
   line, update a product quantity, set a serial number, ...)

With those steps, the point 2. can cause multiple RPC. To limit that, this commit
idea is:
- First, do step 1. for all barcodes;
- Do step 2. only once and for ALL parsed data:
  - Get from the cache already fetched data;
  - Make a RPC to fetch remaining data.
- Then, for each barcode, process it (step 3.) one after each other.

While step 3. can also generate some RPC, it still important to let the
application handles barcode one at the time because we can't predict
repercussions each barcode can have on the followings. But for fetching
data, it is not important to keep the scanned order and it is more
interessting to fetch all needed data in one single RPC.

Update the ongoing barcode count
================================

Let say the user scans 100 barcodes at once, before this commit, the app
will be blocked displaying a message saying "Processing 100 barcodes"
and the app will be unlocked after all 100 barcodes were processed.

With this commit, the behavior is the same but the barcodes count won't
be "100" but will start at "0/100" and will be updated every time a
barcode is processed ("1/100", then "2/100" and so and so.)

Also, a side effect of this change is the app is visually updated every
time a barcode is processed. Before this commit, the app wasn't visually
updated until all barcodes were processed.

The purpose of this change is to let the user knows things still goes
right and the app didn't freezed.

Keep only parse in try/catch
============================

Before this commit, a lot more was done in the `BarcodeModel`
`_parseBarcode` try/catch. Now, only the actual parsing will be done.
The motivation for this change is that the try/catch can hide error not
caused by the parsing itself but simply by wrong code.
So, keep the least code lines possible in the try/catch to avoid to mask
actual error.

task-4089638

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>