[IMP] project_timesheet_forecast{_sale}: improve perf of Timesheets/Planning analysis report
Before this commit, the subquery provided in _number_of_days_without_weekend()
was executed multiple times in _select(), which was decreasing the performance by a lot.
After this commit, we build a CTE to ensure that we only execute this subquery once and
use its result every time it's needed.

task-3902302

closes odoo/enterprise#67943

X-original-commit: ce482901cd159dc51e6a85358008cd675672cc8d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>