[IMP] various: create iap_service data
With the improvement of the iap account model made to the community
codebase, iap_service data has to be created for modules that require it
and tests mocking the model had to be changed to the new
structure. Additionaly the assertQueryCount had to be increased for some
tests, as the iap.account is set outside of _contact_iap, which is mocked
in those tests.

task-3741280

closes odoo/enterprise#67883

Related: odoo/odoo#175562
Related: odoo/upgrade#6346
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>