[IMP] l10n_ke_edi_oscu_pos: sends pos order to etims
This commit add a new way to deal with eTIms in the pos application. Before this commit, the only way of sending sales to etims was to generate the invoice while creating the pos order, then deal with the invoice later on.

Now, everything is automated in the pos order directly, that means when the user post a pos order, she's directly send to etims and all the information returned are displayed on the receipt and the order view.

In case the eTims process got an error, the user can still send the order to eTims from the pos_order view.

Stock moves are also send at the same time thanks to the l10n_ke_edi_oscu_stock module.

task-3876855

closes odoo/enterprise#76247

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>