[FIX] hr_holidays: reset time off allocation name
Steps to reproduce the bug:

- Go to Time Off -> Allocation -> Create
- put a name
- change the allocation type
-> the name is reset

Expected result:
The name should not be reset to the old value

Reason:
The displayed name is relies on the private_name
field to get its value. And thus every time
an onchange is triggered the name is reset.

Fix:
Put the private_name field in the form view
instead of the name field. Because we rely on
the onchange of name on the dashboard to

task-3713656

closes odoo/enterprise#67931

Related: odoo/odoo#153044
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Co-authored-by: maor-odoo <maor@odoo.com>
Co-authored-by: Flotchet <mafl@odoo.com>