[IMP] l10n_mx_edi_website_sale: payment way many2one field
This commit adds the field `l10n_mx_edi_payment_method_id` to the
website checkout, so that the customer can choose the correct option to
be saved into the sale order and to fill the attribute in the XML.

Purpose:
In Mexican electronic invoicing, some fields are mandatory for signing
an invoice. Now, on e-commerce, one field is missing to avoid requests
to cancel invoices due to lack of information. This missing field is
"Payment way" (l10n_mx_edi_payment_method_id).

This commit also restructures the placement of `t-if`/`div`s so that
they appear in the right places and not generate empty divs.

closes odoo/enterprise#68902

Task-id: 4064069
Signed-off-by: Laurent Smet (las) <las@odoo.com>