[FIX] account_reports: fix P&L group by error
Problem:
An issue arises due to a negative analytical distribution. If the
analytical percentage is negative, its sign will match the sign of the
amount_currency of its move line. However, in the query that groups by
the analytical account for P&L, the amount value is inverted. This does
not cause a problem when the analytic percentage is positive because the
amount_currency sign is opposite to the sign of the analytic line amount.
But, in cases where the sign is the same, there will be a balance with a
different sign than the amount, which will not comply with the
constraint of the account.move.line table:
"account_move_line_check_amount_currency_balance_sign".

Steps to reproduce:
1. Activate "Analytic Accounting" in the settings
2. In an invoice, set an analytic distribution with a negative percentage
3. Validate the invoice
4. Go to P&L report an try to group by an analytic account
5. A validation error will occur.

Proposed solution:
Remove the constraint in the temporary table.

opw-4092475
opw-4092232
opw-4086515
opw-4084850
opw-4083077

closes odoo/enterprise#68057

X-original-commit: 02d37544b56f527dce3fca38d1b61f22c521e136
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>