[FIX] l10n_mx_{edi,reports}: xml declaration
This commit refactors the generation of pretty printed raw xml string
to a helper function in `l10n_mx_edi.document` for `l10n_mx_edi` and
`l10n_mx_reports`.

It modifies the xml declaration to be a custom generated string, which
only difference from the normal one is that it uses double quoted string
rather than single quoted string in it.

This way, the top file of the generated xml string will look like this:

```xml
<?xml version="1.0" encoding="UTF-8"?>
```

This ensures as much support as possible from third party vendors
because some do not consider single quoted string to be valid (even
though it should be).

related opw-4085399

closes odoo/enterprise#68214

Task-id: no task
X-original-commit: 608322a722c6910628b674b64064fa27a770b15d
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>