[IMP] spreadsheet_edition: add trending line for odoo chart
Task Description

This task adds the possibility to display trending
lines for odoo bar/line charts. As we are not able to
customize each data series (like in o-spreadsheet),
we have here a global options to allows the trending
line for each data series with the same parameters.

Related Task

closes odoo/enterprise#67969

Task: 3997500
Related: odoo/odoo#175366
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>