[FIX] marketing_automation: removing pager
Current behaviour:
When you create a campaign, if you add more activity (56)
than the default pager limit (40)
The next pager page will be empty (instead of showing
the remaining 16 activities)

Steps to reproduce:
1. Go to Marketing Automation
2. Create a campaign
3. Create more than 40 activities
4. Alternate parents/child when creating
5. The pager will appear, with at least 2 pages
6. Go to the second page
7. Nothing displayed
NB: When inputting the number of activities in
the pager, everything is rendered fine.

Cause of the issue:
To be displayed, the activities are hierarchised, by finding
child and parents (_getSortedRecordsByHierarchy)
However this system works great when all the records are given.
Using a pager, some records are not used in the hierarchy,
causing the outputted list to be incomplete.
(ie if an activity A is in the first 40 and the child in the
last 16, it will not be rendered correctly)

Fix:
Disabling the pager by setting a higher limit

opw-4006610

closes odoo/enterprise#69641

X-original-commit: 4ee7c42dc795ffd9a9d65a9a118f33a88c5cc489
Signed-off-by: Antoine Demany (ande) <ande@odoo.com>