[FIX] l10n_co_reports: only lines with a withholding tax in "Certificado de Retención en IVA"
Steps to reproduce:
- Install Accounting and l10n_co
- Switch to a Colombian company (e.g. CO Company)
- Create a bill with a line having an IVA tax and a withholding tax:
  * IVA Compra 19%
  * RteIVA 15% sobre el 19% IVA
- Post the bill
- Create another bill with a line having only an IVA tax
- Post the bill
- Go to "Accounting / Reporting / Colombian Statements / Certificado de Retención en IVA"

Issue:
"Monto del Pago Sujeto Retención" column is incorrect.
It takes into account the tax amount from the IVA tax of both bills.
As there is no withholding tax on the second bill, it should be excluded from
the report.

Cause:
The report is checking the tax lines where the account is "2408%%" to compute
the amount of this column.
So every tax lines with from an IVA tax will be taken into account, even if
there is no wihholding tax.

Solution:
Computing the correct amount by checking the account and the presence of a
withholding tax is pretty complex.
It would require to rely on the invoice lines instead of the tax lines to
recompute the tax amounts because the amount of a specific tax is aggregated
into a unique tax line (even if there are several invoice lines with different
taxes combination) and there is no information about the tax repartition in the
tax line.
Recomputing all from the invoice lines would bad for the performance.

As the withholding taxes are always 15% over VAT in Colombian localization,
we can simply apply this formula to retrieve the amount subjected to withholding:

[Monto del Pago Sujeto Retención = Retenido Consignado / 15%]

opw-3989661

closes odoo/enterprise#68776

X-original-commit: 8777a6d80eb468173dc0bacad135b87d1eec5b7c
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>