[FIX] website_appointment: manage appointment button styling
Issue
-----
When adding formating to an appointment button linked to a specific
type of appointment, the button will lead to all appointments instead
of just the specific type.

Change
-----
Since buttons can have custom formatting, this will add HTML elements
inside the button element. Therefore, we need to take into account that
the target of the event may not be the button itself.

opw-4080512

closes odoo/enterprise#68037

Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>