[FIX] knowledge: only load child articles once in the sidebar
Purpose:
-------
Currently, clicking several times quickly on the caret to unfold an
article in the sidebar may throw an error and show the child articles
several times in the sidebar.

This commit fixes the issue by preventing to make a new request to load
the child articles if one has already been made.

Task-4099133

closes odoo/enterprise#68748

X-original-commit: e8f11df26b590414875786cd307f7a7277181aa4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Adrien Schoffeniels (adsc) <adsc@odoo.com>