[FIX] knowledge: only load child articles once in the sidebar
Purpose:
-------
Currently, clicking several times quickly on the caret to unfold an
article in the sidebar may throw an error and show the child articles
several times in the sidebar.

This commit fixes the issue by preventing to make a new request to load
the child articles if one has already been made.

Task-4099133

closes odoo/enterprise#68013

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>