[IMP] sale_subscription: redesign of labels journal items
During this commit: https://github.com/odoo/odoo/pull/174988/commits/7fe72283f2380f08eef970f73bebe6877dcda5e3
We changed a bit the _compute_name by adding a new depends on the ref, which was
causing issue for the test_auto_payment_across_time. Since we are writing on the
ref in the _subscription_post_success_payment function, the compute was
triggered again and so the line name was wrong.

The solution we chose is to override the compute and when the line has a sale
order line that is from a recurring invoice, we don't compute it again.

closes odoo/enterprise#67611

Task: 4058751
Related: odoo/odoo#174988
Related: odoo/upgrade#6350
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>