[FIX] hr_attendance_gantt: non deterministic test
The hoot helpers make the time advance more quickly. In AttendanceGanttModel,
open ended records have their dateStopField set to the current moment.
So, depending on the way time passes, the pills can have different ending
columns. Here we remove a useless call to selectGanttRange, so that less
time passes and we are sure we stay between 18h and 19h and not after
19h, so that "Attendance 6" ends at 6pm 02 November 2018 (included).

Build 66529031.

closes odoo/enterprise#68064

X-original-commit: 67390831f78e8b0a0b3a8625297269639c89a705
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>