[FIX] stock_barcode_mrp_subcontracting: dont use default ML src location
**Current behavior:**
When adding an unplanned move line to a transfer in barcode, if
the line is added to an existing move which is a subcontract
transfer, it will not take the subcontract location as its
source location value. Rather, it will use whatever value is
defined on the picking.

**Expected behavior:**
Use the more precise location defined on the move rather than
the picking.

**Steps to reproduce:**
1. Create a product with a subcontract BOM

2. Create a receipt for that product from the subcontractor

3. Open the transfer in Barcode and manually add another line
     for the same product -> validate

4. In the receipt form, look at the move lines (detailed
     operations) and group by location to see the discrepancy

**Cause of the issue:**
When creating a move line here, it does not initially get a
`move_id` value. Later when the move is written to the line,
we would not generally want to override whatever value is on the
line already because it (normally) is more precise/was purposely
changed later than what value is on the move.

**Fix:**
In the case of lines which are created and then linked to a
subcontract move, take the `location_id` of the move rather than
whatever value is on the MoveLine.

opw-4007423

closes odoo/enterprise#69096

X-original-commit: 64797416cc56cc01d0525ebf7f56c4de010ca622
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>