[FIX] sale_subscription: reset env company before posting invoice
To reproduce the bug:

1. Install the subscription feature on your website.
2. Create a second company and link it to the website.
3. Create a recurring product with the service type.
4. In the subscription app, create a quotation with the specified product.
5. Preview the invoice for the quotation.
6. Copy the link for the portal invoice and open it in an incognito window.
7. Pay the invoice.

You will encounter a TraceBack error depending on the modules installed. This is due to the `company_id` being set to an
empty record as a result of how it is computed here:

https://github.com/odoo/odoo/blob/718567eb720cbf66abc9f9096da6fc6ea1aab103/addons/account/models/account_move.py#L696-L698

In our case, `move.journal_id.company_id` (the invoice company) exists, so `self.env.company` (the website company) gets
ignored. However, since the public user of the website doesn't have access to the invoice company, the `company_id` is
set to an empty record after checking the access rights with `._accessible_branches()[:1]`.

opw-3999546

closes odoo/enterprise#68034

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>