[FIX] l10n_au_hr_payroll: fix PERIODS_PER_YEAR definition
PERIODS_PER_YEAR is using "yearly" but hr_contract uses "anually"
and since this model uses the hr_contract, we need to keep the
same dictionary keys.

In customer database the installation of the module is failing
because cannot find key "anually". The issue would also happen
if you just set a yearly wage

opw-4083058

closes odoo/enterprise#68036

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>