[FIX] account_asset: account selectable multicompany in asset
To reproduce:
Be in multicompany, with several companies selected.
Create an asset
Click on the Expense Account field
Select accounts from a different company than the asset's

We should not show these.
Even worse, in ulterior version, you can save the asset.
(computing depreciations would still be prevented)

The issue is that we define a custom domain, so it should include it.

no-task

closes odoo/enterprise#70419

X-original-commit: 6ae78d483177d5bed85563e303c92e4c3076295a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>