[FIX] stock_barcode: fix main menu dark background
Before this commit, the background of the main menu of barcode was using
only the light themed background image. This caused an issue in dark mode,
with the entire UI being dark, while the background was still white. Now,
a div element wrapping the whole content of this menu, with the appropriate
'o_home_menu_background' classname.

This allow the UI to always display the right image, without having to handle
custom lines of scss for this purpose.

closes odoo/enterprise#68075

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>