[FIX] web_studio: fix kanban editor scale
This commit fixes the scale of the kanban editor, which was increased compared
to the size of the standard kanban view. This was made in an attempt to ease the
edition inside of the card, but has become unecessary, since hooks displayed when
dragging an item have already an increased padding to make the drop target larger.

The paddings of the hooks have been adapted for this purpose instead.

task-4075367

Part-of: odoo/enterprise#68084
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>