[IMP] iot: Add print start/finish notifications
Display a notification to the user when a document starts printing and
again when the printing process is complete. Each notification include
the name of the printer (and IoT box name) being used, so the user is
informed about the status of their print job and the specific printer
involved. This enhances the overall experience by providing clear
feedback.

Currently, the user is only notified when the print job is failed.

closes odoo/enterprise#68457

X-original-commit: 83b0385ccc7b971f1cf018a4424fcdd71b806f8b
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>