[IMP] room: add pwa support
Purpose:
--------
Add pwa support for the room booking view so that it can be displayed
in full screen mode on tablets.

In the sidebar of the frontend view, there is now a button "install
application" that opens the pwa installation window.
This button is not shown in the PWA or if the browser does not support
PWAs.
Each room may have its own pwa.

Task-3691870

closes odoo/enterprise#54627

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>