[FIX] industry_fsm: show address on partner field before saving record
steps to reproduce:
-install field service app
-Go to the Field Service app and create a new task.
-Assign a partner (customer) to the task.
-Notice that the address of the assigned partner is not displayed immediately.
-The address only appears after the record is saved.

cause:
-context 'show_address' was not passed to _compute_display_name

Technical:
In a previous pull request https://github.com/odoo/enterprise/pull/56179,
the show_address context was removed.As a result, the
_compute_display_name method was not able to correctly display
the partner's address when a partner was assigned to a task.
The address would only appear after the task was saved.

Solution:
The show_address context should be reintroduced and passed to the
_compute_display_name method during the task creation or update
process. This will allow the address to be displayed immediately
when the partner is selected.

task-4038335

closes odoo/enterprise#68198

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>