[FIX] room: only send presences to interested users
In the community counterpart of this PR, a debounce is added
to the addition of bus channels in order to batch them as much
as possible.

The room module uses the `waitForSubscribe` bus helper but never
awaits it which trigger errors in other tests.

This PR fixes this issue.

closes odoo/enterprise#68408

X-original-commit: 93e260fb403b049bac197c7ab06f4bcc27bc860d
Related: odoo/odoo#176689
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>