[IMP] timesheet_grid: fix focus of start/stop button in timesheet grid header
Before this commit when we try to get the focus off the start button and stop
button and tried pressing ENTER key to start or stop the timer. It works fine as
it is handled by onClick.But all key press are not taken into consideration(key
strokes recorded in input or text-area fields). During the step in onboarding
tour of using Stop Button we are out of focus of stop_timer button as we have
clicked on input field before this step. In tour every button has events both
click and Enter but the target must the source of the element(given in tour
trigger).

After this commit the start and stop button will be in focus when no modal is
displayed or the element focused is not an input and textarea.

task-3997245

closes odoo/enterprise#68357

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>