[FIX] l10n_in_hr_payroll: fix deprecated tz
Asia/Calcutta was removed in ubuntu Noble in favor of Asia/Kolkata

This is only visible in the pure noble build and not in the standard
build. The assumption is that since requirements are installed as root,
the tz is added when installing the pip package.

Since we will soon install requirements as odoo on runbot in the docker
image, this should will start to break in standard builds too.

closes odoo/enterprise#68172

X-original-commit: 836ec62c49dc5d6dfc408ee3b7e8518336ba5e2e
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>