[FIX] room: prevent notification when editing a booking
Purpose:
-------
Currently when editing a booking from the frontend form view, a warning
notification is shown, indicating that the booking being edited has been
updated or deleted. 

This notification should only be shown when the booking being edited has
been updated or deleted by another user (because in that case the user
will be forced to leave the edition view without doing anything).

Task-4104748

closes odoo/enterprise#69139

X-original-commit: 8af2b9027fe4ec30a1be27c36769c5188613dbe6
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>