[FIX] account_reports: datetime-date comparison error
Description of the issue this commit addresses:

When trying to do a period comparison on a report with annotations, a traceback
would be raised because of a comparison between a datetime and a date.

Desired behavior after this commit is merged:

No traceback is raised.

Note on the fix:

In some cases the period_date_from gets a .date() applied which it should not
so this is removed.

task-4045835

closes odoo/enterprise#68197

X-original-commit: ffdab8c0035dcb8d505ab6f267176b08a047aad8
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>