[FIX] test_website_sale_full: adapt test .badge selector
This commit aims to adapt the test to find the right DOM element.

As the front-end badges recently got their design reviewed, we need to
set some badges to `bg-*` (aka old badge design) instead of `text-bg-*`
depending on the context.

As we need to apply that `bg-primary` on the wishlist badge, the test
has to be adapted.

task-4106164

closes odoo/enterprise#68203

Related: odoo/odoo#176254
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>