[IMP] stock_barcode: count product tracked at delivery
Before this commit, it was expected to scan a serial number for quants
tracked by SN with multiple quantity.

This commit changes that. A quant for a product tracked by SN with
multiple quantity means we don't want to set the SN right away in our
stock but only when the product leaves our stock (set the SN at the
delivery for example.)
So, in the Barcode app, while doing an inventory adjustment, the serial
number won't be expected to be scanned anymore for those quants, and the
tracking number element won't be displayed neither on the line.

Also, avoid to mark a inventory line as faulty when quantity to count is
hidden, since it let the user to know what the quantity to count is
(once the line is not red, that would mean it is the supposed quantity
in stock.)

task-3790171

closes odoo/enterprise#62413

Signed-off-by: Steve Van Essche <svs@odoo.com>
Co-authored-by: svs-odoo <svs@odoo.com>