[FIX] account_reports: ensure unique line_ids on accounting report lines
The aim of this commit is to add an extra security in the test of the reporting framework.

Context:
Those line_id are used to recognize lines in the framework and also in the frontend by
owl as t-key. Those line_id not being unique would result in a crash in the
front-end framework (only from 17.0 as OWL is used starting version 17.0 on accounting
reports).

Before the commit:
There isn't any test performed on the line_id property.

After the commit:
assertLinesValues will make sure that each line_id is unique for the generated data.

task-4063612

closes odoo/enterprise#70676

X-original-commit: 50dd60b2ce88e12fa5944728907db369144677bf
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Co-authored-by: Brice Bartoletti <bib@odoo.com>