[FIX] account_reports: ensure unique line_ids on accounting report lines
The aim of this commit is to add an extra security in the test of the reporting framework.

Context:
Those line_id are used to recognize lines in the framework and also in the frontend by
owl as t-key. Those line_id not being unique would result in a crash in the
front-end framework (only from 17.0 as OWL is used starting version 17.0 on accounting
reports).

Before the commit:
There isn't any test performed on the line_id property.

After the commit:
assertLinesValues will make sure that each line_id is unique for the generated data.

Back-ports the following commits:
 - https://github.com/odoo/enterprise/commit/fa3f03d859a35595990f34606c0864f9868e13fc
 - https://github.com/odoo/enterprise/commit/74000d6d5c8435bcaad93f7baf143abdf8593064
 - https://github.com/odoo/enterprise/commit/fed41d1b3042414823345d995eb173f7acb3c6dc

task-4063612

closes odoo/enterprise#70420

X-original-commit: dff1dd7daa6826336ea13bcd98b762f74330f2b8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Co-authored-by: Brice Bartoletti <bib@odoo.com>