[FIX] document: audit trail with document versionning
Normally, when using document versioning, the attachments linked to a
record are re-attached to the document when changing the version.
This is not compatible with the requirement for audit trail in Germany
(GoBD) as we want to be sure that the document won't be deleted.
This is ensure by making sure all the attachments stay attached, and
can't be deleted.

To reproduce:
* Install Accounting and Documents, as well as l10n_de
* Select/configure a German company
* Create new document folder settings for Customer Invoices
* Create a new invoice
* Send & Print
* Delete it from the attachments (it should "detach it")
* Send & Print again

opw-4095549

closes odoo/enterprise#74945

X-original-commit: a5319f56d77b5b81856b9b0ddb1a478eec968cd6
Related: odoo/odoo#189247
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>