[FIX] l10n_ar_edi: fields lenght on Renewal Request.
Version:

17, master

Description of the issue/feature this PR addresses:

It is necessary to restrict the fields lenght that are used to create Renewal Request to argentinean goverment. For example: It is received an error and not downloaded "Renewal Request" file if the company name lenght has more than 64 characters or if the company has a Contact with state name or city name lenght more than 128 characters.

Video showing how to replicate the error:
https://drive.google.com/file/d/1JB39Njjfpf9nJExyixaBLSvcL19Ig49X/view

Steps to reproduce:

1) Log in with admin on runbot odoo enterprise 17 instance and install l10n_ar_edi (Argentinean Electronic Invoicing) module.
2) Take position on company "Responsable Inscripto".
3) Change company name to a name with lenght more than 64 characters.
4) Go to "Accounting / Configuration / Settings" and click on "Generate Renewal Request" on "Argentinean Localization" section. It will be received this message:

"Internal Server Error
The server encountered an internal error and was unable to complete your request. Either the server is overloaded or there is an error in the application."

Current behavior before PR:
It is received an error and not downloaded "Renewal Request" file if the company name lenght has more than 64 characters or if the company has a Contact with state name or city name lenght more than 128 characters.

Desired behavior after PR is merged:
It is downloaded "Renewal Request" file if the company name lenght has more than 64 characters or if the company has a Contact with state name or city name lenght more than 128 characters.

Ticket Adhoc side: 78650

Task latam:

closes odoo/enterprise#69563

X-original-commit: a48931fe2727abffa7af24389c9557d7e1a53e09
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>