[FIX] social_twitter: resolve error on repost, like
Before this Fix
===============
The repost and like operations on the feed view lead to failure.

Reason
======
Due to the modifications in https://github.com/odoo/odoo/blob/
844ef1895c8fd4f39c5bebe7e3cffa5a6e68fe0d/addons/web/controllers/home.py#L37
any RPC calls with URLs starting with 'social_twitter/' are now being redirected
to the mentioned controller, causing errors and operational failures.

After this Fix
==============
Rpcs call will be redirected to the respected controller and operations
will not fail.

Task-4072724

closes odoo/enterprise#69110

X-original-commit: 7bc481894a6c841ee7a14a600c0a173c11938c53
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>