[FIX] product_barcodelookup: missing image traceback
Issue: Traceback is encountered when scanning a barcode that is available in the
barcodelookup api but doesn't have image. Example barcode: 9780349410159

- Traceback: `AttributeError: 'dict' obiect has no attribute 'content'`

This is because of the inconsistent return value of `barcode_lookup_request`. If
it succeeds, it returns a `Response` object but if not, it returns a `dict`.

To fix the traceback, we check if the fetching of image failed by checking if
the returned value is a `dict`. If so, we just skip the image assignment for the
product.

closes odoo/enterprise#68309

Signed-off-by: Fabien Pinckaers (fp) <fp@odoo.com>