[IMP] appointment,*: add calendar synchronization onboarding
*appointment_google_calendar, appointment_microsoft_calendar

For both connectors (google calendar and microsoft calendar),
add a button in an information banner on appointment type form
view to connect their account when all the following conditions
are fulfilled:

- The general sync credentials are set in the global settings
- The sync is not paused in the global settings
- The user had no sync token (meaning they have not sync'd their
calendar yet, or have reset their sync account) This means that
when a user has manually paused their sync, the banner will not
appear, as the idea here is rather to inform they can connect their
account for the first time.

To do so, add AppointmentSyncButton component in appointment module,
that can be parametrized with icon src and calendar name. This
prevents duplication and handle the logic when pressing the button.
They are added in the onboarding banner xml in their respective module.
The component can be overriden to add specific sync data in sub modules

The default use case would be to be redirected in a new tab to google
calendar / microsoft calendar account selection pages. In the unlikely
case the configuration has changed, a notification will be shown instead

[ADD] A small bridge module is added to handle this feature for
microsoft calendar between appointment and microsoft_calendar.

Task-3644621

Part-of: odoo/enterprise#66162
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Clément Cardot (cacl) <cacl@odoo.com>
Co-authored-by: Noé Antoine (nan) <nan@odoo.com>