[FIX] quality_control_worksheet: prevent error when validating quality worksheets template
The error occurs when we validate the quality worksheet template.

Steps to reproduce:
- Install the 'quality_mrp_workorder_worksheet' module
- Quality > Configuration > Quality Worksheet Templates
- Click on 'Design Template'
- Click on 'Close'
- Click on 'Validate'

Traceback: ValueError
Expected singleton : quality.check()

This error occurs at [1], where we are receiving 'self' as empty.

Previous behaviour:
When we try to design a template in the 'Quality Worksheet Templates' the
studio opens as expected. However, if we close it without making any changes,
we are redirected to the form view. In this view, we encounter strange behavior
and are unable to perform any actions.

After behaviour:
Now, when we click on 'Design Template' from the list view and then close it,
we are redirected to the form view of the 'Quality Worksheet Templates.'

[1]- https://github.com/odoo/enterprise/blob/7eb1131b413afc2549382d2524ec83445803b682/quality_mrp_workorder_worksheet/models/quality.py#L12

sentry-5695518176

closes odoo/enterprise#72797

X-original-commit: a3841880f635e2f72eded0f7d3d134cea1cb4cb3
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>