[IMP] account_reports,l10n_eu_oss_report: periodicity for OSS reports
OSS reports may need a different periodicity than the one configured in the tax report.
Another periodicity specifically for the OSS report must be present when the module is installed so that the closing can get the periodicity
related to the report. This also includes a way to compute the tax closing for the OSS report as there were no tax closing for these

upgrade: https://github.com/odoo/upgrade/pull/6390
task-4110459

Part-of: odoo/enterprise#68551
Related: odoo/odoo#180095
Related: odoo/upgrade#6390
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>