[FIX] account_accountant: Avoid SQL Error on Uninstalling through res.config.settings
This commit uses the built-in functionality of the execute() function by creating the field module_sign. This will trigger installation of the sign module under the same circumstances while avoiding any registry errors that are caused by updated fields on res.comapny. Otherwise, when uninstalling a module through res.config settings (i.e. unchecking the Partner Autocomplete box), the old method would throw an error if the uninstalled module had any fields on res.company. When using the related field, it tries to access a cached copy of the res.company record with field values that may have been modified, as opposed to using the native functionality of execute to install a missing module.

closes odoo/enterprise#74000

Opw: 4107361
X-original-commit: 8faed5815fbd7b30c5d93dd2f0b6cfda163a3c56
Signed-off-by: William André (wan) <wan@odoo.com>