[FIX] website_sale_renting: fix weekend product unavailable tour
The Issue:
Prior to this commit, starting a tour on Thursday and setting
the end date to Saturday or Sunday would cause a traceback error.
This occurred because renting was not allowed on weekends,
leading to the tour's failure.

The Fix:
The fix involved enabling availability for renting on Saturdays and Sundays.

runbot-72624

closes odoo/enterprise#68380

X-original-commit: 0025e58134097f0a31e83da4d1bfa63daa73b3ea
Signed-off-by: Amzil Ayoub (amay) <amay@odoo.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>