[FIX] room: only send presences to interested users
In the community counterpart of this PR, a debounce is added
to the addition of bus channels in order to batch them as much
as possible.

The room module uses the `waitForSubscribe` bus helper but never
awaits it which trigger errors in other tests.

This PR fixes this issue.

backport of https://github.com/odoo/enterprise/pull/68099

closes odoo/enterprise#68361

Related: odoo/odoo#176580
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>