[FIX] account_budget: fix the computation of is_over_budget in purchase order
Before this commit:
- When updating the quantity or unit price on a purchase order line, red lines
would appear even if it was not overcommitted.
- The budget amount was calculated by adding the committed amount from the budget
line and the subtotal of the line item multiplied by the (ordered quantity - invoiced quantity).
- If the invoiced quantity is 0, the calculation would multiply the ordered
quantity by the line item’s subtotal, leading to an incorrect calculation.

After this commit:
- The is_over_budget check will be calculated based on the latest budget line
committed amount by using flush_model in the `_compute_all` method.

Task-4112959

closes odoo/enterprise#68526

X-original-commit: a360e782460c655a72f2f39ab712fe7752168eb9
Signed-off-by: William André (wan) <wan@odoo.com>