[FIX] helpdesk: merge duplicate help tiket sla
Coupled with 53d6c3492ec977e4588ee734f3c3a7f06534687a

Steps to reproduce:
- Helpdesk > Configuration > SLA policies
- Create one if necessary or use existing
- Helpdesk > Tickets > Create
- Configure it so that it corresponds to SLA requirements
- Duplicate the ticket and merge them both (data_merge module required)

Currently helpdesk ticket uses the default merge fuction, which simply
concatenates the values of merged records. This causes a few strange
things to happen:

1. sla_ids which registers SLA policy types, has duplicate entries. This
recordset should remain a set.

2. Followers of the merged tickets are sometimes not added to the
resulting record. The default merge simply swaps the resource id with no
regard for a possible duplication of unique key (res_model, res_id,
partner_id), which causes the transaction to fail on repeat followers.
This is a common problem with every record inheriting mail.thread which
does not define its own merge function as well.

3. Keeping every SLA status record is redundant, merging tickets makes
it impossbile to keep track of which task each deadline was originally
related to anyway. Hence why we only keep those with the shortest
deadline for each type. Priority is not a concern here since SLA
policies are recomputed according to a ticket's priority rating, so we
cannot have 2 SLAs with the same policy and different priorities.

opw-4037262

closes odoo/enterprise#70946

X-original-commit: e00c03e7a94cd1a4002f7c6812c31cf650c5db46
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>