[FIX] account_Reports: traceback on reports export with annotations
Description of the issue this commit addresses:

When trying to export a report in .xlsx format, if some annotations have been
added to the report, and no growth comparison is set, a traceback appears.

Steps to reproduce:

1. Install Accounting
2. Go to any report
3. Add an annotation on any line
4. Export the report in .xlsx format
5. A traceback appears

Desired behavior after this commit is merged:

Doing the same steps after this fix should result in the export's success and
the possibility for the user to open the file in which the annotations are.

Note on the fix:

During investigation, it was noticed that the offset has a discordant
computation behaviors. After completing a subheader column, the offset is
incremented by the colspan but not after an option column is completed.
This means that the condition that was computing the annotations' column
offset would not have been required if the offset was correct in the first
place. Correcting that computation makes the whole system clearer.

Also some tests have been added to make sure the behavior of the annotation
offset stays constant.

opw-4095385

closes odoo/enterprise#68388

X-original-commit: 538e8ca3da75a65573d3ac19735e93b25f1bd75b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>