[FIX] spreadsheet_edition: add case for duplicating on comment plugin
Steps to reproduce:
- create a new spreadsheet
- duplicate the first sheet
- add any columns in the second sheet

Duplicating a sheet would not add the sheet ID to the cellThreads map in
the cell threads plugin. This would cause issues when adding rows or
columns because the sheetId can't be referenced and an error is thrown.
Added a case in handle in order to add the new sheet id from duplicating
to the cellThreads map.

opw-4100746

closes odoo/enterprise#68922

X-original-commit: 8510ed129017bd0e1de56e3613701427aff89309
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>