[IMP] industry_fsm: disable retry on TestIndustryFsmUi
FSM does really strange stuff, as a result when the tour fails the
retry tends to completely lose its shit and generate a different and
useless error. See [runbot build 66859325] as an example, the actual
failure is the step "Switch to kanban view" not finding its kanban
button, but the retries try to re-run with a dynamically generated
table which has likely been rolled back (?).

Disable retry on the test_ui tour for now, maybe there's a team in
charge of FSM which could handle this situation properly tho...

[runbot build 66859325]: https://runbot.odoo.com/runbot/build/66859325

closes odoo/enterprise#68420

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>