[FIX] account_reports: tour failing with variant and sections
The tour `test_sections_tour` fails undeterministically.
It fails after opening the the variant, then opening another
section of it.

The issue is that we try to open the second section just before
the preloading of this section happens.
The opening of it enters `loadReportOptions`, has no loadingCallNumber
as it is the first opening.
We then stop at the call to `get_options`.

Then, the preloading can happen (which is why it's undeterministic).
We enter the `if` in `preLoadClosedSections`, as the options are not
yet of the second section. We then increment loadingCallNumber, which
means it's not in sync with the call that was made to the python.
If the options call comes back now, we don't remake a call with the new
loadingCallNumber.

runbot-75429
runbot-74795

closes odoo/enterprise#68910

X-original-commit: ab39d19f7ad9088bf9a75f7edef72e22938b115d
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>