[FIX] account_reports: properly handle hide_if_zero option and blank_if_zero columns
Partial revert of https://github.com/odoo/enterprise/commit/318ae33da2e36036e1a74028a36d02ad4fa5964e (we keep the test and revert the fix).

The reverted commit intended had hide_if_zero consider string values as 0 so that report lines could be hidden if the only non-0 values on
them were strings (like customer name, VAT number, ...). However, in 17.2+, this caused issues with the blank_if_zero columns, which then
never showed such string values.

We revert the original fix and take a new approach in this commit. We do that in 17.0 instead of 17.2 just for homogeneity of the code base,
and ease of maintenance on longer term. The new strategy is bacisally to consider non-number values as 0 only when handling the hide_if_zero option,
without impacting to the computation of the is_zero key of column dicts.

Also, this commit explicitly adds the 'percentage' figure_type to the ones checked by the 'is_zero' key in column dicts: percentages are numeric values;
ignoring them there was wrong.

closes odoo/enterprise#68524

X-original-commit: 51ca2649e854ce7f848fa766d4528076ff3363aa
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>