[FIX] hr_payroll: handle multiple salary lines with same type
Issue
-----
Error when multiple input lines on a payslip are of the same type.

Steps
-----
[hr_payroll]
1. Create a salary attachment for an employee, type "Attachment of salary",
with a monthly amount and total amount A.
2. Create another salary attachment for the same employee, same type and monthly
amount and total amount B different from A.
3. Create a payslip for the employee, create a contract with a start date matching
the salary attachment date. On "Other Inputs", remove the "Attachment of salary"
line. Create two input lines of type "Attachment of salary", one with amount A and
another with amount B.
4. Compute sheet > Confirm > Mark as paid > ** Error **

Cause
-----
Generally, input lines of the same type on a payslip will be merged in one input line,
but it is not the case if there are multiple salary attachments of the same type matching
these input lines.

opw-4066851

closes odoo/enterprise#68996

X-original-commit: 17518be4536fbc858d57bd457eee5b32087826a6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>