[FIX] test_mail_enterprise: test whatsapp webpush notification
Performs test of a related commit in communiy:

[FIX] mail: include whatsapp in webpush notifications
There are couple of components in the notification
architecture. For the sake of illustration
let's simplify it with the following: Sender -> Queuer -> Receiver
Queuer is an FMC backend, so we should only be concerned about the
other two. Receiver which is just a frontend service worker that will
generate notifications as soon as he receives an event. The issue
(before this commit) was in the Sender part, which was skipping
whatsapp type message generation.

[Reproduce]
- Configure Whatsapp account, include user A in "Notify users"
- Allow notifications in your browser (chat icon/ OdooBot suggestion)
- Send a whatsapp message to the configured whatsapp number
- BUG: Push notification not showing up

opw-3720699

closes odoo/enterprise#68453

Related: odoo/odoo#175638
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>