[FIX] sign: cog menu signature request fail in new records
Before this commit, when a new record was just created, requesting a
signature from the cog menu would not call the Signature Request action
since we were fetching the record id from the wrong place (controller
props).

After this commit, when creating a new record, Signature request action
is sucessfully called in newly created records since we fetch the record
id from the controller state instead, which is mutable to record changes.

task-4116177

closes odoo/enterprise#69365

X-original-commit: fe6eca2ea31d58b5c6a2f01c73c429540c0d93ef
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>