[FIX] industry_fsm_report: add a check to correctly save worksheet template in tour
Before this commit, when the step on the `fsm_task_form_tour` tour saves
the worksheet template creation, the save action will take some times to
be able to generate a new worksheet model for that worksheet template.
The problem is the tour does not wait enough time before doing the next
steps.

This commit adds a step to wait the form view dialog of the worksheet
template is closed before doing the other steps.

runbot-72641

closes odoo/enterprise#68714

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>